+++ docs/source/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.10.9 (2025-08-21)
+- allow setting S3 endpoint URL in pipedag.yaml
 - Fix: IBM DB2 works with colspec
 
 ## 0.10.8 (2025-08-18)